style(windows): silence MSVC C4305/C4244 conversion warnings in F2Dock sources
Mechanically apply C-style casts to silence MSVC double-to-float (C4305)
and double-to-int / size_t conversion warnings (C4244) reported by the
Windows Debug CI job. Casts use the existing F2Dock idiom `(T)(expr)`,
matching patterns already present in the codebase (e.g. `(int)ceil(...)`).

Changes:
- inc/f2dock/ElementInformation.h: append `f` suffix to float-field
  literals in the element-property table (310 sites).
- src/f2dock/Docking.cpp: wrap `ceil(...)`/`floor(...)` int-target
  initializers and Matrix::translation / Vector / retrieveTransformation
  arguments (110 sites).
- src/fft-utils/fastfft.cpp: cast `sizeof(...)*N` to `size_t` for
  FFTW_malloc, cast grid dims to int (33 sites).
- src/PostFiltering/FilterOutput.cpp (16), src/fast-clash/clashFilter.cpp
  (9), src/fast-PQ/PairingHeap.cpp (9), src/fast-resCont/resContFilter.cpp
  (9), src/GB-rerank/pairingHeap.cpp (9), src/f2dock/F2Dock.cpp (9),
  src/math/Matrix.cpp (2), src/f2dock/TopValues.cpp (1).

Of the ~678 warning sites in F2Dock-owned code, 517 are addressed by
this scripted pass. The remaining ~161 sites are skipped — they involve
multi-line expressions, dependent-type STL container `.size()` returns,
SSE intrinsic helpers, and complex PG-range arithmetic that need manual
review. They will be tackled separately.

libcvc warnings (`build/_deps/libcvc-src/...`) are deliberately not
touched in this PR and will be addressed in libcvc upstream.

Linux/macOS GCC and Clang do not emit C4305/C4244, so no functional
change is expected on those platforms; the casts are equivalent to the
implicit conversions they replace.

+++ src/GB-rerank/pairingHeap.cpp
@@ -23,6 +23,8 @@
 
 #include "GB-rerank/pairingHeap.h"
 
+#include <cstdint>
+
 int PairingHeap::new_node(int x, VAL_TYPE k) {
   int fp;
 
@@ -57,9 +59,9 @@ void PairingHeap::free_node(int xp) {
 
 void PairingHeap::add_to_aux_area(int xp) {
   if (auxptr == NIL)
-    min_auxptr = auxptr = (*heap)[xp + LEFT] = (*heap)[xp + RIGHT] = xp;
+    min_auxptr = auxptr = (*heap)[xp + LEFT] = (int)((*heap)[xp + RIGHT] = xp);
   else {
-    int yp = (*heap)[auxptr + LEFT];
+    int yp = (int)((*heap)[auxptr + LEFT]);
 
     (*heap)[xp + LEFT] = yp;
     (*heap)[xp + RIGHT] = auxptr;
@@ -88,7 +90,7 @@ inline int PairingHeap::pairing_heap_link(int xp, int yp) {
   (*heap)[yp + LEFT] = xp;
   (*heap)[yp + RIGHT] = (*heap)[xp + CHILD];
   if ((*heap)[yp + RIGHT] != NIL)
-    (*heap)[(*heap)[yp + RIGHT] + LEFT] = yp;
+    (*heap)[(uint64_t)((*heap)[yp + RIGHT] + LEFT)] = yp;
   (*heap)[xp + CHILD] = yp;
 
   (*heap)[xp + LEFT] = (*heap)[xp + RIGHT] = NIL;
@@ -221,7 +223,7 @@ void PairingHeap::Delete_Min(int &mx, VAL_TYPE &mk) {
     mk = INF;
   } else {
     if (use_aux_trees && (auxptr != NIL)) {
-      (*heap)[(*heap)[auxptr + LEFT] + RIGHT] = NIL;
+      (*heap)[(uint64_t)((*heap)[auxptr + LEFT] + RIGHT)] = NIL;
       (*heap)[auxptr + LEFT] = NIL;
 
       auxptr = multi_pass_merge(auxptr);
@@ -233,7 +235,7 @@ void PairingHeap::Delete_Min(int &mx, VAL_TYPE &mk) {
     mx = (*heap)[root + ID];
     mk = (*heap)[root + KEY];
 
-    int xp = (*heap)[root + CHILD];
+    int xp = (int)((*heap)[root + CHILD]);
 
     free_node(root);
 
@@ -252,12 +254,12 @@ void PairingHeap::Decrease_Key(int xp, VAL_TYPE k) {
 
     if (xp != root) {
       if ((*heap)[xp + RIGHT] != NIL)
-        (*heap)[(*heap)[xp + RIGHT] + LEFT] = (*heap)[xp + LEFT];
+        (*heap)[(uint64_t)((*heap)[xp + RIGHT] + LEFT)] = (*heap)[xp + LEFT];
 
-      if ((*heap)[(*heap)[xp + LEFT] + CHILD] == xp)
-        (*heap)[(*heap)[xp + LEFT] + CHILD] = (*heap)[xp + RIGHT];
+      if ((*heap)[(uint64_t)((*heap)[xp + LEFT] + CHILD)] == xp)
+        (*heap)[(uint64_t)((*heap)[xp + LEFT] + CHILD)] = (*heap)[xp + RIGHT];
       else
-        (*heap)[(*heap)[xp + LEFT] + RIGHT] = (*heap)[xp + RIGHT];
+        (*heap)[(uint64_t)((*heap)[xp + LEFT] + RIGHT)] = (*heap)[xp + RIGHT];
 
       (*heap)[xp + LEFT] = (*heap)[xp + RIGHT] = NIL;
