+++ src/PostFiltering/FilterOutput.cpp
@@ -161,8 +161,10 @@ void applyFilter(clashFilter *fFilter,
     int nclashes = 0, nclose = 0;
     Matrix4x4 m = transformations.at(i);
     CCVOpenGLMath::Matrix transM(
-        m(0, 0), m(0, 1), m(0, 2), m(0, 3), m(1, 0), m(1, 1), m(1, 2), m(1, 3),
-        m(2, 0), m(2, 1), m(2, 2), m(2, 3), m(3, 0), m(3, 1), m(3, 2), m(3, 3));
+        (float)(m(0, 0)), (float)(m(0, 1)), (float)(m(0, 2)), (float)(m(0, 3)),
+        (float)(m(1, 0)), (float)(m(1, 1)), (float)(m(1, 2)), (float)(m(1, 3)),
+        (float)(m(2, 0)), (float)(m(2, 1)), (float)(m(2, 2)), (float)(m(2, 3)),
+        (float)(m(3, 0)), (float)(m(3, 1)), (float)(m(3, 2)), (float)(m(3, 3)));
 
     fFilter->computeInteractions(transM, &nclashes, &nclose, &intVal);
